+++ doc/release-notes.md
@@ -34,6 +34,8 @@ This part of documentation collects descriptive release notes to capture the mai
 
 **Minor Changes and bug-fixing**
 
+* Add a check for existence of a cutout file when retrieving cutout [PR #1846](https://github.com/pypsa-meets-earth/pypsa-earth/pull/1846)
+
 * Avoid crash in `build_osm_network` when merging substations with the same `station_id` if `tag_substation` or `symbol` contains missing OSM metadata [PR #1818](https://github.com/pypsa-meets-earth/pypsa-earth/pull/1818)
 
 * Add CI to update reference objective values [PR #1811](https://github.com/pypsa-meets-earth/pypsa-earth/pull/1811)